+++ vite-plugin-ssr/node/runtime/renderPage/getPageAssets/getManifestEntry.ts
@@ -1,6 +1,6 @@
 export { getManifestEntry }
 
-import type { ViteManifest, ViteManifestEntry } from '../../../helpers'
+import type { ViteManifest, ViteManifestEntry } from '../../helpers'
 import { assert, slice, isNpmPackageModule } from '../../../utils'
 import { assertClientEntryId } from './assertClientEntryId'
