+++ vite-plugin-ssr/node/runtime/renderPage/getPageAssets/retrieveAssetsProd.ts
@@ -1,7 +1,7 @@
 export { retrieveAssetsProd }
 
 import { assert, isNpmPackageModule } from '../../../utils'
-import { ViteManifest } from '../../../helpers'
+import { ViteManifest } from '../../helpers'
 import { getManifestEntry } from './getManifestEntry'
 import { extractAssetsAddQuery } from '../../../plugin/plugins/extractAssetsPlugin/extractAssetsAddQuery'
 import type { ClientDependency } from '../../../../shared/getPageFiles/analyzePageClientSide/ClientDependency'
